Optimize class name and configuration file for other-feature-example project
