+++ examples/shardingsphere-jdbc-example/other-feature-example/encrypt-example/encrypt-spring-boot-example/src/main/java/org/apache/shardingsphere/example/encrypt/table/spring/boot/EncryptSpringBootExample.java
@@ -31,10 +31,10 @@ import java.sql.SQLException;
 @ComponentScan("org.apache.shardingsphere.example.core.jpa")
 @EntityScan(basePackages = "org.apache.shardingsphere.example.core.jpa.entity")
 @SpringBootApplication(exclude = JtaAutoConfiguration.class)
-public class ExampleMain {
+public class EncryptSpringBootExample {
     
     public static void main(final String[] args) throws SQLException {
-        try (ConfigurableApplicationContext applicationContext = SpringApplication.run(ExampleMain.class, args)) {
+        try (ConfigurableApplicationContext applicationContext = SpringApplication.run(EncryptSpringBootExample.class, args)) {
             ExampleExecuteTemplate.run(applicationContext.getBean("encryptExample", ExampleService.class));
         }
     }
