+++ examples/shardingsphere-jdbc-example/other-feature-example/hint-example/hint-raw-jdbc-example/src/main/java/org/apache/shardingsphere/example/hint/raw/jdbc/HintRawExample.java
@@ -29,8 +29,8 @@ import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.Statement;
 
-public final class ExampleMain {
-    
+public final class HintRawExample {
+
     private static final HintType TYPE = HintType.DATABASE_TABLES;
 //    private static final HintType TYPE = HintType.DATABASE_ONLY;
 //    private static final HintType TYPE = HintType.WRITE_ONLY;
@@ -57,7 +57,7 @@ public final class ExampleMain {
     }
     
     private static File getFile(final String configFile) {
-        return new File(ExampleMain.class.getResource(configFile).getFile());
+        return new File(HintRawExample.class.getResource(configFile).getFile());
     }
     
     private static ExampleService getExampleService(final DataSource dataSource) {
